requirements.txt
tkinter
Pillow==9.0.0
pyinstaller==6.6.0
